+++ src/main/java/com/messenger/myperfectmessenger/controller/UserController.java
@@ -2,6 +2,7 @@ package com.messenger.myperfectmessenger.controller;
 
 import com.messenger.myperfectmessenger.domain.User;
 import com.messenger.myperfectmessenger.exception.UserNotFoundException;
+import com.messenger.myperfectmessenger.security.service.SecurityService;
 import com.messenger.myperfectmessenger.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,24 +16,31 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.security.Principal;
 import java.util.List;
 
 @RestController //для REST архитектуры
 @RequestMapping("/user")
 public class UserController {
     private final UserService userService;
+    private final SecurityService securityService;
 
-    public UserController(UserService userService) {
+    public UserController(UserService userService, SecurityService securityService) {
         this.userService = userService;
+        this.securityService = securityService;
     }
 
     @GetMapping
-    public ResponseEntity<List<User>> getUsers() {
-        List<User> users = userService.getUsers();
-        if (users.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        } else {
-            return new ResponseEntity<>(users, HttpStatus.OK);
+    public ResponseEntity<List<User>> getUsers(Principal principal) {
+        if(securityService.checkIfAdmin(principal.getName())){
+            List<User> users = userService.getUsers(principal);
+            if (users.isEmpty()) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            } else {
+                return new ResponseEntity<>(users, HttpStatus.OK);
+            }
+        } else{
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
@@ -49,26 +57,40 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUser(@PathVariable Long id) {
-        User user = userService.getUser(id).orElseThrow(UserNotFoundException::new);
-        return new ResponseEntity<>(user, HttpStatus.OK);
+    public ResponseEntity<User> getUser(@PathVariable Long id, Principal principal) {
+        User user = userService.getUser(id, principal);
+        if(user != null){
+            return new ResponseEntity<>(user, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
-
+    //maybe should be deleted(or only for admins)
     @PostMapping
     public ResponseEntity<HttpStatus> createUser(@RequestBody User user) {
         userService.createUser(user);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
+
     @PutMapping
-    public ResponseEntity<HttpStatus> updateUser(@RequestBody User user) {
-        userService.updateUser(user);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<HttpStatus> updateUser(@RequestBody User user, Principal principal) {
+        if(securityService.checkIfAdmin(principal.getName()) || (securityService.getUserIdByLogin(principal.getName())==user.getId())){
+            userService.updateUser(user);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteUser(@PathVariable Long id) {
-        userService.deleteUserById(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<HttpStatus> deleteUser(@PathVariable Long id, Principal principal) {
+        if(securityService.checkIfAdmin(principal.getName()) || (securityService.getUserIdByLogin(principal.getName())==id)){
+            userService.deleteUserById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
+
 }
